Fixed bug in microsurface code more properly.
The microsurface node struct has a field "child_index" that wasn't large
enough to address all of the nodes when there were more than one time sample.

For now, I'm just making that field larger.  In the future this can be
dealt with better by using e.g. an implicit indexing scheme.

+++ config.hpp
@@ -7,7 +7,7 @@ namespace Config
 {
 extern float dice_rate;
 extern float min_upoly_size;
-extern size_t max_grid_size;
+extern uint8_t max_grid_size;
 extern float grid_cache_size;
 
 extern int bucket_size;
